test : 로직 수정에 따른 테스트 수정
